[UPDATE] trying to fix code duplication
