+++ modules/users/change_password/common_functions.py
@@ -1,6 +1,21 @@
 import boto3
 import json
 from botocore.exceptions import ClientError
+import hashlib
+import hmac
+import base64
+
+
+def get_secret_hash(username, client_id, client_secret):
+    message = username + client_id
+    dig = hmac.new(client_secret.encode('utf-8'), message.encode('utf-8'), hashlib.sha256).digest()
+    return base64.b64encode(dig).decode()
+
+
+def get_secret():
+    secret_name = "users_pool/client_secret"
+    secret = get_secret_value(secret_name)
+    return secret
 
 
 def get_secret_value(secret_name, region_name="us-east-2"):
